+++ modal/stub.py
@@ -470,6 +470,7 @@ class _Stub:
         max_inputs: Optional[
             int
         ] = None,  # Limits the number of inputs a container handles before shutting down. Use `max_inputs = 1` for single-use containers.
+        _experimental_boost: bool = False,  # Experimental flag for lower latency function execution (alpha).
     ) -> Callable[..., _Function]:
         """Decorator to register a new Modal function with this stub."""
         if isinstance(_warn_parentheses_missing, _Image):
@@ -550,6 +551,7 @@ class _Stub:
                 allow_background_volume_commits=_allow_background_volume_commits,
                 block_network=block_network,
                 max_inputs=max_inputs,
+                _experimental_boost=_experimental_boost,
             )
 
             self._add_function(function)
@@ -590,6 +592,7 @@ class _Stub:
         max_inputs: Optional[
             int
         ] = None,  # Limits the number of inputs a container handles before shutting down. Use `max_inputs = 1` for single-use containers.
+        _experimental_boost: bool = False,  # Experimental flag for lower latency function execution (alpha).
     ) -> Callable[[CLS_T], _Cls]:
         if _warn_parentheses_missing:
             raise InvalidError("Did you forget parentheses? Suggestion: `@stub.cls()`.")
@@ -620,6 +623,7 @@ class _Stub:
             block_network=block_network,
             _allow_background_volume_commits=_allow_background_volume_commits,
             max_inputs=max_inputs,
+            _experimental_boost=_experimental_boost,
         )
 
         def wrapper(user_cls: CLS_T) -> _Cls: